fix: add .int() to trainingExperienceYears; increase AI insights timeout to 25s
trainingExperienceYears accepted 2.7 or similar fractional values — should
always be a whole integer. Consistent with the .int() pattern on other
experience-type fields across the codebase.

AIInsightsCard: 12s abort timeout was too short for AI calls on slow
networks or queued backends. Increased to 25s to match realistic AI
response times.

+++ server/src/routes/user.ts
@@ -109,7 +109,7 @@ router.patch('/profile', authenticate, async (req: AuthRequest, res: Response) =
       fitnessLevel: z.string().transform(v => v.toUpperCase()).pipe(
         z.enum(['BEGINNER', 'INTERMEDIATE', 'ADVANCED', 'EXPERT'])
       ).optional(),
-      trainingExperienceYears: z.number().finite().min(0).max(80).optional(),
+      trainingExperienceYears: z.number().int().finite().min(0).max(80).optional(),
       avatarUrl: z.string().url('Некорректный URL').max(2048).refine((u) => u.startsWith('https://'), 'URL должен использовать HTTPS').optional(),
     });
